ek updated to v5 + TravisCI config updated
