+++ .travis.yml
@@ -1,11 +1,24 @@
+sudo: false
+
 language: go
 
 go:
-  - 1.5.3
-  - 1.6
+  - 1.6.x
+  - 1.7.x
   - tip
 
-sudo: false
+branches:
+  only:
+  - master
+  - develop
+
+os:
+  - linux
+
+matrix:
+  fast_finish: true
+  allow_failures:
+    - go: tip
 
 before_install:
   - make deps
